requirements.txt
streamlit>=1.28.0
cryptography>=41.0.0
vaderSentiment>=3.3.2
openai>=1.0.0
python-dotenv>=1.0.0
